fix: how secrets are passed to workflows (part 2)
I thought that jobs.<job_id>.secrets.<secret_id> might support the
'default' key like jobs.<job_id>.inputs.<input_id> does. An
example (ref:1) contained a 'description', which is not documented in
the former's specs, but *is* for the latter, which I took to imply a
possibility. An error put that assumption to rest (ref:2).

Amend: bffb40487ec8
Ref: https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#example-2
Ref: https://github.com/doomemacs/themes/actions/runs/2268993373